+++ packages/ui/src/components/dotted-spinner/dotted-spinner.tsx
@@ -6,7 +6,7 @@ import {
   useStyleConfig,
   HTMLChakraProps,
 } from "@chakra-ui/react";
-import { cx } from "@companydotcom/utils";
+import { cx } from "@vastly/utils";
 // TODO: Remove the workaround whenever MS fixes the issue
 // https://github.com/microsoft/TypeScript/issues/48212
 import type { ComponentWithAs } from "@chakra-ui/react";