+++ apps/workshop/CHANGELOG.md
@@ -1,5 +1,37 @@
 # workshop
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @vastly/forms@0.4.0
+  - @vastly/ui@0.4.0
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @vastly/forms@0.3.0
+  - @vastly/ui@0.3.0
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [2b140ed]
+  - @vastly/forms@0.2.0
+  - @vastly/ui@0.2.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [1ccfdc7]
+  - @vastly/forms@0.1.0
+  - @vastly/ui@0.1.0
+
 ## 0.0.2
 
 ### Patch Changes